fix: correct eventDefaultThumbnail require paths for relocated assets
Metro resolves require() at runtime, so these escaped tsc during the
layout port.
