+++ src/social/features/events/EventDetail/Components/EventCoverImage.tsx
@@ -2,7 +2,7 @@ import { memo, useState } from 'react';
 import { Image, StyleSheet, View } from 'react-native';
 import { FileRepository } from '@amityco/ts-sdk-react-native';
 
-const eventDefaultThumbnail = require('../../../../assets/images/eventDefaultThumbnail.png');
+const eventDefaultThumbnail = require('../../../../../core/assets/images/eventDefaultThumbnail.png');
 
 const styles = StyleSheet.create({
   cover: {
